Fix link errors with newer compilers
Without this change, ICU 5.0 uses the C++11 type char16_t in its C++
APIs. This requires client programs to be compiled with the same
options (eg -std=c++11 for g++) which typically enables the whole of
C++11 and is currently undesirable.

The problem doesn't occur with g++ versions less than 4.7 because they
don't have -std=c++11. However, 4.7 has now arrived in Debian wheezy
and Ubuntu quantal.

This change is a cherry-pick of r32780 from upstream svn.